# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Eplus v 0.0.1",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/educational-experiences/{educationalExperience}/groups": {
            "get": {
                "summary": "Mostrar todos los grupos de una experiencia educativa",
                "operationId": "b5e4446b612bdd03080c5fdd0632a860",
                "parameters": [
                    {
                        "name": "educationalExperience",
                        "in": "path",
                        "description": "ID de la experiencia educativa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los grupos de la experiencia educativa"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "post": {
                "summary": "Crear un nuevo grupo en una experiencia educativa",
                "operationId": "4992efc8b25236456681b587f2adcbe3",
                "parameters": [
                    {
                        "name": "educationalExperience",
                        "in": "path",
                        "description": "ID de la experiencia educativa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nombre del grupo",
                                        "type": "string"
                                    },
                                    "shift": {
                                        "description": "Turno del grupo",
                                        "type": "string"
                                    },
                                    "period": {
                                        "description": "Periodo del grupo",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Grupo creado exitosamente"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
        "/api/educational-experiences/{educationalExperience}/groups/create": {
            "get": {
                "summary": "Mostrar los detalles de la experiencia educativa para crear un grupo",
                "operationId": "80cfbcd563de0ac06ae3e7d94d5740ce",
                "parameters": [
                    {
                        "name": "educationalExperience",
                        "in": "path",
                        "description": "ID de la experiencia educativa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la experiencia educativa"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
        "/api/educational-experiences/{educationalExperience}/groups/{groupId}": {
            "get": {
                "summary": "Mostrar un grupo por su ID",
                "operationId": "7c16f51e7d29761601cf7e70bdd519cd",
                "parameters": [
                    {
                        "name": "educationalExperience",
                        "in": "path",
                        "description": "ID de la experiencia educativa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "ID del grupo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del grupo"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "put": {
                "summary": "Actualizar un grupo",
                "operationId": "11452bb48a3045dd84a7aa9fd0795d3b",
                "parameters": [
                    {
                        "name": "educationalExperience",
                        "in": "path",
                        "description": "ID de la experiencia educativa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "ID del grupo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nombre del grupo",
                                        "type": "string"
                                    },
                                    "shift": {
                                        "description": "Turno del grupo",
                                        "type": "string"
                                    },
                                    "period": {
                                        "description": "Periodo del grupo",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Grupo actualizado exitosamente"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar un grupo",
                "operationId": "ef61c07a3b21eec54da394cb79b4e657",
                "parameters": [
                    {
                        "name": "educationalExperience",
                        "in": "path",
                        "description": "ID de la experiencia educativa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "ID del grupo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Grupo eliminado exitosamente"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
        "/api/educational-experiences/{educationalExperience}/groups/{groupId}/students": {
            "get": {
                "summary": "Obtener los estudiantes de un grupo",
                "operationId": "6dea73df9abb169b29314a8865942610",
                "parameters": [
                    {
                        "name": "educationalExperience",
                        "in": "path",
                        "description": "ID de la experiencia educativa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "ID del grupo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de estudiantes en el grupo"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
...
    }
}